requirements.txt
pandas~=1.2.5
gspread~=3.7.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
